services/interaction-service/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
pydantic
pika
pyjwt[crypto]
prometheus-client